refactor: move highlight copy into the renderers module
Address #365 review: the build*Sentence functions (and their helpers /
copy maps) now live in session-highlight-renderers.tsx alongside the
markup that uses them, rather than being imported from the model. The
model is now pure data — no copy, no date-fns formatting.
